requirements.txt
django
djangorestframework
environs
Pillow